identifiers: Rename room_or_room_alias_id module to match the type
