+++ crates/ruma-common/src/identifiers.rs
@@ -29,7 +29,7 @@ pub use self::{
     mxc_uri::{MxcUri, OwnedMxcUri},
     room_alias_id::{OwnedRoomAliasId, RoomAliasId},
     room_id::{OwnedRoomId, RoomId},
-    room_or_room_alias_id::{OwnedRoomOrAliasId, RoomOrAliasId},
+    room_or_alias_id::{OwnedRoomOrAliasId, RoomOrAliasId},
     room_version_id::RoomVersionId,
     server_name::{OwnedServerName, ServerName},
     session_id::{OwnedSessionId, SessionId},
@@ -53,7 +53,7 @@ mod key_name;
 mod mxc_uri;
 mod room_alias_id;
 mod room_id;
-mod room_or_room_alias_id;
+mod room_or_alias_id;
 mod room_version_id;
 mod server_name;
 mod session_id;
