+++ crates/ruma-common/src/identifiers/room_or_alias_id.rs
